Marching cubes implemented on top of voxel base. Needs fine tuning (deformation is dropping some inputs right now).

+++ CGD2019.P1.MiningGame/Assets/scripts/Voxel.cs
@@ -9,15 +9,15 @@ namespace App.Gameplay {
     [Serializable]
     public class Voxel {
 
-        private byte type;
+        private int type;
         public VoxelType Type { get { return (VoxelType)type; } }
 
         public Voxel() {
-
+            
         }
 
         public Voxel(VoxelType type) {
-            this.type = (byte)type;
+            this.type = (int)type;
         }
     }
 }
